+++ src/gamspy/__init__.py
@@ -44,6 +44,7 @@ from gamspy._model import Model
 from gamspy._model import ModelStatus
 from gamspy._model import Problem
 from gamspy._model import Sense
+from gamspy._options import ModelInstanceOptions
 from gamspy._options import Options
 from gamspy._symbols import Alias
 from gamspy._symbols import Equation
@@ -79,6 +80,7 @@ __all__ = [
     "Ord",
     "Card",
     "Options",
+    "ModelInstanceOptions",
     "EngineConfig",
     "NeosClient",
     "SpecialValues",